+++ src/index.js
@@ -1,6 +1,0 @@
-module.exports = {
-  toTs: require('./to-ts'),
-  toMd: require('./to-md'),
-  toMock: require('./to-mock'),
-  mockServer: require('./mock-server'),
-};