+++ src/util/index.ts
@@ -1,0 +1,3 @@
+import { AriesConfig } from "../interface";
+
+export const defineConfig = (config: AriesConfig) => config